pyproject.toml
[tool.poetry]
name = "equestria"
version = "0.1"
description = "A web-based collection of forced-alignment tools"
authors = ["GiPHouse <unknown@unknown.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.3"
pillow = "^7.0.0"
requests = "^2.23.0"
black = "^19.10b0"
clam = "^3.0.11"
coverage = "^5.0.4"
selenium = "^3.141.0"
django-nose = "^1.4.6"
codecov = "^2.0.22"
django-background-tasks = "^1.2.5"
selenium-wire = "^1.0.11"
django-nested-inline = "^0.4.2"
django-guardian = "^2.2.0"
psycopg2-binary = "^2.8.5"
uwsgi = "^2.0.18"
django-colorfield = "^0.3.0"
djangorestframework = "^3.12.2"


[tool.poetry.dev-dependencies]
pydocstyle = "^5.0"
black = {version = "^19.10b0"}

[tool.black]
line-length = 80
target-version = ["py37"]
exclude = '''
/(
  migrations
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

